+++ app/src/main/java/com/pixel/caffeine/presentation/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.pixel.caffeine.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.pixel.caffeine.presentation.theme.CaffeineApp
+import com.pixel.caffeine.presentation.theme.CaffeineTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            CaffeineTheme {
+                CaffeineApp()
+            }
+        }
+    }
+}
