+++ app/src/main/java/com/pixel/caffeine/presentation/theme/Font.kt
@@ -1,4 +1,4 @@
-package com.pixel.caffeine.ui.theme
+package com.pixel.caffeine.presentation.theme
 
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
